[mle] fix bug in leader active/pending dataset synchronization (#5021)
The leader is the authoritative source when propagating active/pending
operational datasets. When the leader reboots, it is possible for the
leader to become out-of-sync. In particular, after the leader resets,
it restores its network data and active/pending operational datasets
using MLE Data Request/Response to retrieve them from a neighboring
device. If the neighboring device has an older active/pending
operational dataset, the leader will retrieve the older datasets and
never attempt to propagate the newer datasets that it may have stored
locally.

This commit makes the following changes:

- The leader does not accept any changes to the active/pending
  operational datasets.

- After retrieving the latest network data from a neighboring device
  (as part of resynchronizing after reset), the leader will increment
  the full and stable network data versions to ensure that the leader
  propagates the latest active/pending operational datasets.